pyproject.toml
[tool.poetry]
name = "text-2-beluga-discord-bot-1"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]
readme = "README.md"
packages = [{include = "text_2_beluga_discord_bot_1"}]

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.0.0"
pilmoji = {extras = ["requests"], version = "^2.0.3"}
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
